+++ database/migrations/2022_04_10_150348_create_student_subject_table.php
@@ -14,8 +14,8 @@ return new class extends Migration
     public function up()
     {
         Schema::create('student_subject', function (Blueprint $table) {
-            $table->integer('student_id')->unsigned();
-            $table->integer('subject_id')->unsigned();
+            $table->unsignedBigInteger('student_id')->unsigned();
+            $table->unsignedBigInteger('subject_id')->unsigned();
             $table->float('mark');
             $table->foreign('student_id')->references('id')->on('students');
             $table->foreign('subject_id')->references('id')->on('subjects');
